iio: adc: ad7768-1: Use new spi-engine implementation
Add  support for new spi-engine implementation on ad7768-1 driver.
This optimizes the offload transfers by reducing the CS delay and
removing the number of commands on offload's FIFO, which makes
the low latency mode viable.

Without the otimization there's not enough time between the end of a
transfer and the start of the next one, reducing the actual sample rate.

Signed-off-by: Jonathan Santos <Jonathan.Santos@analog.com>