iio: adc: ad7768-1: Add support for Synchronization over SPI
The synchronization method using GPIO requires the generated pulse to be
truly synchronous with the base MCLK signal. When it is not possible to
do that in hardware, the datasheet recommends using synchronization over
SPI, where the generated pulse is already synchronous with MCLK.

This requires the SYNC_OUT pin to be connected to SYNC_IN pin.

Signed-off-by: Jonathan Santos <Jonathan.Santos@analog.com>